UserTextAreaコンポーネントを追加し、SystemTextAreaコンポーネントを削除した
components/SystemTextArea copy.tsx
- ファイルを削除

components/UserTextArea.tsx
- useChatGPTフックからinputとsetInputを受け取る
- キーボードのEnterキーとCtrlキーまたはMetaキーが押されたときに、requestChatGPT関数を呼び出す
- textarea要素を返す

hooks/useChatGPT.tsx
- inputとsetInputをエクスポート

pages/index.tsx
- UserTextAreaコンポーネントをインポート
- SystemTextAreaコンポーネントを削除
- UserTextAreaコンポーネントをSystemTextAreaコンポーネントの代わりに表示するように変更
